Splatoon 3 compatibility fixes
Warning: Splatoon 3 and 2 are incompatible by design, that is, the clear
screen key from 2 is used to exit the post in 3 so there is no way that
the same functionality can be used in both. Unless, that is, someone
changes how this program works.

The following are changes to make this post printer work in Splatoon 3.
The clear of the screen is now done via the left stick button and the
initial position of the pencils has to be brought to the left-most
selection.

Everything else remains the same. We are lucky that the canvas stayed
320x120p and that the printing method also persisted from Splatoon 2.
This is why some minor fixes can make this printer work in Splatoon 3.

Signed-off-by: Tar XF <tarxf@esquibits.com>